+++ lib/CalDAV/Schedule/Plugin.php
@@ -3,6 +3,7 @@
 namespace Sabre\CalDAV\Schedule;
 
 use
+    DateTimeZone,
     Sabre\DAV\Server,
     Sabre\DAV\ServerPlugin,
     Sabre\DAV\Property\Href,
@@ -790,13 +791,23 @@ class Plugin extends ServerPlugin {
             }
 
             $sct = $caldavNS . 'schedule-calendar-transp';
-            $props = $node->getProperties([$sct]);
+            $ctz = $caldavNS . 'calendar-timezone';
+            $props = $node->getProperties([$sct, $ctz]);
 
             if (isset($props[$sct]) && $props[$sct]->getValue() == ScheduleCalendarTransp::TRANSPARENT) {
+                // If a calendar is marked as 'transparent', it means we must
+                // ignore it for free-busy purposes.
                 continue;
             }
 
             $aclPlugin->checkPrivileges($homeSet . $node->getName() ,$caldavNS . 'read-free-busy');
+
+            if (isset($props[$ctz])) {
+                $vtimezoneObj = VObject\Reader::read($props[$ctz]);
+                $calendarTimeZone = $vtimezoneObj->VTIMEZONE->getTimeZone();
+            } else {
+                $calendarTimeZone = new DateTimeZone('UTC');
+            }
 
             // Getting the list of object uris within the time-range
             $urls = $node->calendarQuery([
@@ -834,6 +845,7 @@ class Plugin extends ServerPlugin {
         $generator->setObjects($objects);
         $generator->setTimeRange($start, $end);
         $generator->setBaseObject($vcalendar);
+        $generator->setTimeZone($calendarTimeZone);
 
         $result = $generator->getResult();